display users table

+++ EnvirenmentTravail/src/main/java/ma/youcode/GestionDabsence/DAO/SecretaireDAO/SecretaireDAO.java
@@ -12,12 +12,11 @@ public interface SecretaireDAO {
 
     public List<Secretaire> getAll() throws ClassNotFoundException, SQLException;
 
-    public Secretaire getById(int idSecretaire) throws ClassNotFoundException, SQLException;
+    public Secretaire getById(Long idSecretaire) throws ClassNotFoundException, SQLException;
 
     public Secretaire sauveSecretaire(String nom, String prenom, String numTele, String email, String password, String CIN, String dateNaissance) throws ClassNotFoundException, SQLException;
 
-    public void updateSecretaire(int idSecretaire,String nom, String prenom, String numTele, String email, String password, String CIN, String dateNaissance) throws ClassNotFoundException, SQLException;
-
+    public void updateSecretaire(Long idSecretaire,String nom, String prenom, String numTele, String email, String password, String CIN, String dateNaissance) throws ClassNotFoundException, SQLException;
 
     public void deleteById(int id) throws ClassNotFoundException, SQLException;
